+++ packages/sdk/src/config/index.ts
@@ -11,9 +11,8 @@ const TESTNET_CONFIG: SDKOptions = {
   network: Network.TESTNET,
   contractAddress:
     "0x69faed94da99abb7316cb3ec2eeaa1b961a47349fad8c584f67a930b0d14fec7",
-  hyperionFullNodeIndexerURL: "https://api-testnet.hyperion.xyz/v1/graphql",
+  hyperionFullNodeIndexerURL: "https://api-testnet.hyperion.xyz",
   hyperionAPIHost: "https://api-testnet.hyperion.xyz",
-  officialFullNodeIndexerURL: "https://api.testnet.aptoslabs.com/v1/graphql",
   APTOS_API_KEY: "",
 };
 
@@ -21,9 +20,8 @@ const MAINNET_CONFIG: SDKOptions = {
   network: Network.MAINNET,
   contractAddress:
     "0x8b4a2c4bb53857c718a04c020b98f8c2e1f99a68b0f57389a8bf5434cd22e05c",
-  hyperionFullNodeIndexerURL: "https://api.hyperion.xyz/v1/graphql",
+  hyperionFullNodeIndexerURL: "https://api.hyperion.xyz",
   hyperionAPIHost: "https://api.hyperion.xyz",
-  officialFullNodeIndexerURL: "https://api.mainnet.aptoslabs.com/v1/graphql",
   APTOS_API_KEY: "",
 };